# Conditional Clinger's fast path