As remarked by @jakubjelinek, we cannot unconditionally use the conventional Clinger's fast path because it assumes that the rounding mode is set to 'nearest' (which is the universal default). Now, nothing stops a user from changing the rounding mode, but we need our code to produce the same result (rounding to nearest) irrespective of the system's rounding mode.

Checking that `fegetround() == FE_TONEAREST` is too expensive on some systems.

A better solution was proposed by  @mwalcott3  and involves doing an addition, a subtraction and a comparison to verify that the rounding mode is set to nearest. Because parsing a float might already involve hundreds of instructions, this check is relatively cheap. The resulting branch is also invariably well predicted.

This allows us to bring back the Clinger's fast path when the test is ok (which is always in practice). When it does not apply, we can still use another 'Clinger-like' fast path (based on a proposal by @jakubjelinek).

The result is an increased performance when the Clinger's fast path applies and the 'Clinger-like' fast path would not.

It can cause a slightly performance regression in cases (such as 'canada.txt') when the fast path is not beneficial because we now have a more expensive check, though this may depend on the compiler. There might be ways to micro-optimize this better than what the current PR proposes.

## Ice Lake processor, GCC 11:

Current code:
```
-f data/canada.txt
fastfloat                               :   933.44 MB/s (+/- 1.4 %)    53.64 Mfloat/s  
-f data/canada_short.txt
fastfloat                               :   374.20 MB/s (+/- 1.2 %)    69.53 Mfloat/s  
-f data/mesh.txt
fastfloat                               :   676.27 MB/s (+/- 1.7 %)    92.13 Mfloat/s  
-m uniform
fastfloat                               :  1219.35 MB/s (+/- 2.3 %)    58.12 Mfloat/s  
-m uniform -c
fastfloat                               :   951.23 MB/s (+/- 2.4 %)    54.59 Mfloat/s  
-m simple_uniform32
fastfloat                               :  1225.69 MB/s (+/- 3.0 %)    58.42 Mfloat/s  
-m simple_uniform32 -c
fastfloat                               :   944.60 MB/s (+/- 2.1 %)    54.22 Mfloat/s  
-m simple_int32
fastfloat                               :   771.49 MB/s (+/- 3.1 %)    83.05 Mfloat/s 
```

This PR:
```
-f data/canada.txt
fastfloat                               :   953.92 MB/s (+/- 12.2 %)    54.82 Mfloat/s  
-f data/canada_short.txt
fastfloat                               :   584.63 MB/s (+/- 7.2 %)   108.62 Mfloat/s  
-f data/mesh.txt
fastfloat                               :   809.69 MB/s (+/- 3.6 %)   110.30 Mfloat/s  
-m uniform
fastfloat                               :  1204.92 MB/s (+/- 2.1 %)    57.43 Mfloat/s  
-m uniform -c
fastfloat                               :  1025.67 MB/s (+/- 2.2 %)    58.88 Mfloat/s  
-m simple_uniform32
fastfloat                               :  1222.26 MB/s (+/- 2.0 %)    58.26 Mfloat/s  
-m simple_uniform32 -c
fastfloat                               :  1019.27 MB/s (+/- 5.3 %)    58.49 Mfloat/s  
-m simple_int32
fastfloat                               :   772.84 MB/s (+/- 15.4 %)    83.20 Mfloat/s  
```

Using `fegetround() == FE_TONEAREST` to guard Clinger's (slow):

```
-f data/canada.txt
fastfloat                               :   778.68 MB/s (+/- 1.8 %)    44.75 Mfloat/s  
-f data/canada_short.txt
fastfloat                               :   482.55 MB/s (+/- 1.3 %)    89.66 Mfloat/s  
-f data/mesh.txt
fastfloat                               :   631.72 MB/s (+/- 2.0 %)    86.06 Mfloat/s  
-m uniform
fastfloat                               :   915.72 MB/s (+/- 5.3 %)    43.65 Mfloat/s  
-m uniform -c
fastfloat                               :   992.28 MB/s (+/- 3.0 %)    56.94 Mfloat/s  
-m simple_uniform32
fastfloat                               :   915.51 MB/s (+/- 1.1 %)    43.64 Mfloat/s  
-m simple_uniform32 -c
fastfloat                               :   992.24 MB/s (+/- 3.2 %)    56.95 Mfloat/s  
-m simple_int32
fastfloat                               :   693.57 MB/s (+/- 3.6 %)    74.67 Mfloat/s  
```

## Apple M2 processor, LLVM 14:

current code:

```
-f data/canada.txt
fastfloat                               :  1387.19 MB/s (+/- 0.7 %)    79.72 Mfloat/s      15.90 i/B   290.11 i/f (+/- 0.0 %)      2.41 c/B    43.96 c/f (+/- 0.5 %)      6.60 i/c      3.50 GHz
-f data/canada_short.txt
fastfloat                               :   615.83 MB/s (+/- 1.1 %)   114.42 Mfloat/s      37.59 i/B   212.15 i/f (+/- 0.0 %)      5.28 c/B    29.79 c/f (+/- 0.3 %)      7.12 i/c      3.41 GHz
-f data/mesh.txt
fastfloat                               :   878.58 MB/s (+/- 1.9 %)   119.69 Mfloat/s      24.01 i/B   184.79 i/f (+/- 0.0 %)      3.69 c/B    28.40 c/f (+/- 1.8 %)      6.51 i/c      3.40 GHz
-m uniform
fastfloat                               :  1894.65 MB/s (+/- 0.7 %)    90.30 Mfloat/s      12.59 i/B   277.04 i/f (+/- 0.0 %)      1.72 c/B    37.74 c/f (+/- 0.2 %)      7.34 i/c      3.41 GHz
-m uniform -c
fastfloat                               :  1404.54 MB/s (+/- 0.6 %)    80.61 Mfloat/s      13.08 i/B   239.04 i/f (+/- 0.0 %)      2.31 c/B    42.28 c/f (+/- 0.2 %)      5.65 i/c      3.41 GHz
-m simple_uniform32
fastfloat                               :  1894.15 MB/s (+/- 0.4 %)    90.28 Mfloat/s      12.59 i/B   277.04 i/f (+/- 0.0 %)      1.72 c/B    37.75 c/f (+/- 0.2 %)      7.34 i/c      3.41 GHz
-m simple_uniform32 -c
fastfloat                               :  1414.14 MB/s (+/- 1.4 %)    81.16 Mfloat/s      13.08 i/B   239.01 i/f (+/- 0.0 %)      2.30 c/B    42.00 c/f (+/- 1.0 %)      5.69 i/c      3.41 GHz
-m simple_int32
fastfloat                               :   998.10 MB/s (+/- 0.8 %)   107.46 Mfloat/s      20.17 i/B   196.39 i/f (+/- 0.0 %)      3.26 c/B    31.72 c/f (+/- 0.2 %)      6.19 i/c      3.41 GHz
```

this PR:
```
-f data/canada.txt
fastfloat                               :  1312.35 MB/s (+/- 3.7 %)    75.42 Mfloat/s      16.71 i/B   304.92 i/f (+/- 0.0 %)      2.52 c/B    46.06 c/f (+/- 1.0 %)      6.62 i/c      3.47 GHz
-f data/canada_short.txt
fastfloat                               :   764.14 MB/s (+/- 3.6 %)   141.98 Mfloat/s      30.64 i/B   172.94 i/f (+/- 0.0 %)      4.34 c/B    24.48 c/f (+/- 0.4 %)      7.07 i/c      3.48 GHz
-f data/mesh.txt
fastfloat                               :   964.26 MB/s (+/- 1.3 %)   131.36 Mfloat/s      22.32 i/B   171.83 i/f (+/- 0.0 %)      3.34 c/B    25.72 c/f (+/- 1.8 %)      6.68 i/c      3.38 GHz
-m uniform
fastfloat                               :  1784.08 MB/s (+/- 0.8 %)    85.03 Mfloat/s      13.50 i/B   297.04 i/f (+/- 0.0 %)      1.82 c/B    40.08 c/f (+/- 0.1 %)      7.41 i/c      3.41 GHz
-m uniform -c
fastfloat                               :  1409.95 MB/s (+/- 1.2 %)    80.93 Mfloat/s      12.03 i/B   219.84 i/f (+/- 0.0 %)      2.31 c/B    42.11 c/f (+/- 0.9 %)      5.22 i/c      3.41 GHz
-m simple_uniform32
fastfloat                               :  1783.64 MB/s (+/- 0.4 %)    85.01 Mfloat/s      13.50 i/B   297.04 i/f (+/- 0.0 %)      1.82 c/B    40.09 c/f (+/- 0.2 %)      7.41 i/c      3.41 GHz
-m simple_uniform32 -c
fastfloat                               :  1409.90 MB/s (+/- 1.2 %)    80.91 Mfloat/s      12.03 i/B   219.90 i/f (+/- 0.0 %)      2.31 c/B    42.13 c/f (+/- 0.8 %)      5.22 i/c      3.41 GHz
-m simple_int32
fastfloat                               :   969.15 MB/s (+/- 0.8 %)   104.31 Mfloat/s      21.09 i/B   205.43 i/f (+/- 0.0 %)      3.35 c/B    32.67 c/f (+/- 0.2 %)      6.29 i/c      3.41 GHz
```



cc @jakubjelinek  
credit to @mwalcott3