@lemire 2022-11-16
I think I found an issue in ClangCL + Win32 whereas double(0) becomes -0.0 when fegetround() == FE_DOWNWARD.

I suppose that it could be argued to be correct in some sense.