@jakubjelinek 2022-11-16
Wouldn't it be better to perform most or all of the old/new CLinger's fast path checks first and only if they are all satisfied
call the function to check rounding mode through the floating point ops?  The limited fast path that works in all rounding modes  ought to be a subset of the other.
Of course, only if it improves the benchmarks.