@lemire 2022-11-16
@jakubjelinek My current impression is that given good enough code generation, the `detail::rounds_to_nearest()` call ought to be free. It should always return true, and it is only a handful of instructions, so only a few percent of our instruction count, and there is no data dependency.