@lemire 2022-11-18
It looks like it might be very slightly faster.

Test on an Apple M2 processor (LLVM 14).

Current PR:

```
-f data/canada.txt
fastfloat                               :  1312.83 MB/s (+/- 0.6 %)    75.44 Mfloat/s      16.71 i/B   304.92 i/f (+/- 0.0 %)      2.55 c/B    46.44 c/f (+/- 0.2 %)      6.57 i/c      3.50 GHz 
-f data/canada_short.txt
fastfloat                               :   770.62 MB/s (+/- 0.4 %)   143.18 Mfloat/s      30.64 i/B   172.94 i/f (+/- 0.0 %)      4.34 c/B    24.48 c/f (+/- 0.2 %)      7.07 i/c      3.50 GHz 
-f data/mesh.txt
fastfloat                               :   969.71 MB/s (+/- 1.8 %)   132.10 Mfloat/s      22.32 i/B   171.83 i/f (+/- 0.0 %)      3.35 c/B    25.80 c/f (+/- 1.6 %)      6.66 i/c      3.41 GHz 
-m uniform
fastfloat                               :  1833.26 MB/s (+/- 4.1 %)    87.38 Mfloat/s      13.50 i/B   297.04 i/f (+/- 0.0 %)      1.82 c/B    40.08 c/f (+/- 0.3 %)      7.41 i/c      3.50 GHz 
-m uniform -c
fastfloat                               :  1393.32 MB/s (+/- 1.0 %)    79.97 Mfloat/s      12.04 i/B   219.92 i/f (+/- 0.0 %)      2.33 c/B    42.62 c/f (+/- 0.3 %)      5.16 i/c      3.41 GHz 
-m simple_uniform32
fastfloat                               :  1834.72 MB/s (+/- 3.8 %)    87.45 Mfloat/s      13.50 i/B   297.04 i/f (+/- 0.0 %)      1.82 c/B    40.07 c/f (+/- 0.3 %)      7.41 i/c      3.50 GHz 
-m simple_uniform32 -c
fastfloat                               :  1437.16 MB/s (+/- 3.1 %)    82.49 Mfloat/s      12.04 i/B   219.87 i/f (+/- 0.0 %)      2.31 c/B    42.14 c/f (+/- 1.0 %)      5.22 i/c      3.48 GHz 
-m simple_int32
fastfloat                               :   969.05 MB/s (+/- 1.2 %)   104.30 Mfloat/s      21.09 i/B   205.43 i/f (+/- 0.0 %)      3.35 c/B    32.68 c/f (+/- 0.2 %)      6.29 i/c      3.41 GHz 
```



New proposal:

```
-f data/canada.txt
fastfloat                               :  1370.11 MB/s (+/- 0.3 %)    78.74 Mfloat/s      16.45 i/B   300.18 i/f (+/- 0.0 %)      2.44 c/B    44.50 c/f (+/- 0.2 %)      6.75 i/c      3.50 GHz 
-f data/canada_short.txt
fastfloat                               :   770.58 MB/s (+/- 3.5 %)   143.17 Mfloat/s      30.29 i/B   170.94 i/f (+/- 0.0 %)      4.34 c/B    24.48 c/f (+/- 0.3 %)      6.98 i/c      3.50 GHz 
-f data/mesh.txt
fastfloat                               :   964.19 MB/s (+/- 2.2 %)   131.35 Mfloat/s      22.06 i/B   169.83 i/f (+/- 0.0 %)      3.33 c/B    25.67 c/f (+/- 2.0 %)      6.62 i/c      3.37 GHz 
-m uniform
fastfloat                               :  1829.66 MB/s (+/- 1.5 %)    87.21 Mfloat/s      13.27 i/B   292.04 i/f (+/- 0.0 %)      1.78 c/B    39.08 c/f (+/- 0.2 %)      7.47 i/c      3.41 GHz 
-m uniform -c
fastfloat                               :  1417.53 MB/s (+/- 0.8 %)    81.36 Mfloat/s      11.87 i/B   216.86 i/f (+/- 0.0 %)      2.29 c/B    41.89 c/f (+/- 0.3 %)      5.18 i/c      3.41 GHz 
-m simple_uniform32
fastfloat                               :  1829.92 MB/s (+/- 0.9 %)    87.22 Mfloat/s      13.27 i/B   292.03 i/f (+/- 0.0 %)      1.78 c/B    39.08 c/f (+/- 0.3 %)      7.47 i/c      3.41 GHz 
-m simple_uniform32 -c
fastfloat                               :  1410.67 MB/s (+/- 0.7 %)    80.94 Mfloat/s      11.88 i/B   217.17 i/f (+/- 0.0 %)      2.30 c/B    42.11 c/f (+/- 0.2 %)      5.16 i/c      3.41 GHz 
-m simple_int32
fastfloat                               :   973.45 MB/s (+/- 1.4 %)   104.79 Mfloat/s      20.88 i/B   203.41 i/f (+/- 0.0 %)      3.34 c/B    32.52 c/f (+/- 0.2 %)      6.25 i/c      3.41 GHz 
```

I will test on x64 next.